Edits and minutes
Started working on editing of schedules.
Changed weather modification to be minutes as per issue #2.
Created a class for schedules to be able to define a default value for
some things. This makes it easier when updating to ensure that every
attribute has an value even if the schedule was created in an older
version.
Also implemented a possible fix for issue #12
Corrected an issue in Home that prevented it from showing any schedules
in default mode.